LibWeb: Remove the C++ selector engine
Keep selector traversal, combinator evaluation, pseudo-element
transitions, and structural and functional pseudo-classes entirely in
Rust. Reduce C++ selector matching to synchronous semantic DOM queries
and existing invalidation caches used through FFI.

Pass MatchContext directly across the FFI now that comparison mode no
longer needs to suppress side effects. Drop the matching parameters,
caches, invalidation helpers, and traversal callbacks that only served
the deleted C++ engine.

Rename the remaining integration surface to SelectorMatching and update
all callers. Remove the old C++ matching implementation without a
fallback path.
