+++ Libraries/LibWeb/CSS/Selector.h
@@ -100,7 +100,6 @@ public:
             int step_size { 0 }; // "A"
             int offset = { 0 };  // "B"
 
-            bool matches(int index) const;
             Utf16String serialize() const;
             void serialize_to(Utf16StringBuilder&) const;
         };
@@ -253,7 +252,6 @@ public:
 
     Vector<CompoundSelector> const& compound_selectors() const { return m_compound_selectors; }
     Optional<PseudoElement> target_pseudo_element() const { return m_target_pseudo_element; }
-    bool contains_pseudo_element_transition() const { return m_contains_pseudo_element_transition; }
     NonnullRefPtr<Selector> relative_to(SimpleSelector const&) const;
     bool contains_the_nesting_selector() const { return m_contains_the_nesting_selector; }
     bool contains_pseudo_class(PseudoClass pseudo_class) const { return m_contained_pseudo_classes.get(pseudo_class); }
@@ -265,10 +263,7 @@ public:
 
     auto const& ancestor_hashes() const { return m_ancestor_hashes; }
 
-    bool can_use_fast_matches() const { return m_can_use_fast_matches; }
     bool can_use_ancestor_filter() const { return m_can_use_ancestor_filter; }
-
-    size_t sibling_invalidation_distance() const;
 
     bool is_slotted() const { return m_contains_slotted_pseudo_element; }
     bool has_part_pseudo_element() const { return m_contains_part_pseudo_element; }
@@ -279,19 +274,14 @@ public:
         return *m_rust_selector;
     }
 
-    u64 rust_selector_id() const;
-
 private:
     explicit Selector(Vector<CompoundSelector>&&);
 
     Vector<CompoundSelector> m_compound_selectors;
     mutable Optional<u32> m_specificity;
     Optional<PseudoElement> m_target_pseudo_element;
-    mutable Optional<size_t> m_sibling_invalidation_distance;
-    bool m_can_use_fast_matches { false };
     bool m_can_use_ancestor_filter { false };
     bool m_contains_the_nesting_selector { false };
-    bool m_contains_pseudo_element_transition { false };
     bool m_contains_slotted_pseudo_element { false };
     bool m_contains_part_pseudo_element { false };
